Fix: Bugs

+++ src/components/user/Logout.jsx
@@ -1,20 +1,30 @@
-// For Test
-import { useContext } from "react";
 import Axios from "axios";
 import { useNavigate } from 'react-router-dom';
-import { UserId } from "App";
+import { api } from "config/api";
 
 const Logout = () => { 
-    const {userId, setUserId} = useContext(UserId);
 
     const navigate = useNavigate();
 
     const logout = async () => {
-        Axios.get('http://localhost:4000/users/logout');
-        setUserId(null);
-        localStorage.clear();
-        console.log(userId);
-        navigate('/')
+        const token = localStorage.getItem("loging-token");
+        await Axios.get(`${api.url}/users/logout`, {
+            headers: {
+                "Authorization": token,
+                "withCredentials": true,
+            }
+        })
+        .then((res) => {
+            if(res.status === 200){
+                alert(res.data.message);
+                localStorage.clear();
+                navigate('/');
+            }else{
+                alert("잘못된 요청입니다.")
+            }
+            
+        })
+        
     };
     logout();
     return;
